requirements.txt
requests
python-decouple
